+++ src/Kalnoy/Cruddy/Console/CompileCommand.php
@@ -1,0 +1,77 @@
+<?php
+
+namespace Kalnoy\Cruddy\Console;
+
+use Illuminate\Console\Command;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
+use Kalnoy\Cruddy\Compiler;
+
+class CompileCommand extends Command {
+
+    /**
+     * The filesystem object.
+     *
+     * @var \Illuminate\Filesystem\Filesystem
+     */
+    protected $compiler;
+
+    /**
+     * The console command name.
+     *
+     * @var string
+     */
+    protected $name = 'cruddy:compile';
+
+    /**
+     * The console command description.
+     *
+     * @var string
+     */
+    protected $description = 'Compile a schema for quicker access.';
+
+    /**
+     * Create a new command instance.
+     *
+     * @return void
+     */
+    public function __construct(Compiler $compiler)
+    {
+        parent::__construct();
+
+        $this->compiler = $compiler;
+    }
+
+    /**
+     * Execute the console command.
+     *
+     * @return mixed
+     */
+    public function fire()
+    {
+        $this->compiler->compile($this->argument('locale'));
+    }
+
+    /**
+     * Get the console command arguments.
+     *
+     * @return array
+     */
+    protected function getArguments()
+    {
+        return array(
+            array('locale', InputArgument::OPTIONAL | InputArgument::IS_ARRAY, 'The locale id.'),
+        );
+    }
+
+    /**
+     * Get the console command options.
+     *
+     * @return array
+     */
+    protected function getOptions()
+    {
+        return array();
+    }
+
+}