+++ src/routes.php
@@ -7,42 +7,42 @@ Route::group(['prefix' => Config::get('cruddy::uri')], function ()
         Route::get('_schema',
         [
             'as' => 'cruddy.api.schema',
-            'uses' => 'Kalnoy\Cruddy\EntityApiController@schema',
+            'uses' => 'Kalnoy\Cruddy\Controllers\EntityApiController@schema',
         ]);
 
         Route::get('{entity}',
         [
             'as' => 'cruddy.api.entity.index',
-            'uses' => 'Kalnoy\Cruddy\EntityApiController@index',
+            'uses' => 'Kalnoy\Cruddy\Controllers\EntityApiController@index',
         ]);
 
         Route::post('{entity}',
         [
             'as' => 'cruddy.api.entity.create',
-            'uses' => 'Kalnoy\Cruddy\EntityApiController@create',
+            'uses' => 'Kalnoy\Cruddy\Controllers\EntityApiController@create',
         ]);
 
         Route::get('{entity}/{id}',
         [
             'as' => 'cruddy.api.entity.show',
-            'uses' => 'Kalnoy\Cruddy\EntityApiController@show',
+            'uses' => 'Kalnoy\Cruddy\Controllers\EntityApiController@show',
         ]);
 
         Route::put('{entity}/{id}',
         [
             'as' => 'cruddy.api.entity.update',
-            'uses' => 'Kalnoy\Cruddy\EntityApiController@update',
+            'uses' => 'Kalnoy\Cruddy\Controllers\EntityApiController@update',
         ]);
 
         Route::delete('{entity}/{id}',
         [
             'as' => 'cruddy.api.entity.destroy',
-            'uses' => 'Kalnoy\Cruddy\EntityApiController@destroy',
+            'uses' => 'Kalnoy\Cruddy\Controllers\EntityApiController@destroy',
         ]);
     });
 
-    Route::get('/', 'Kalnoy\Cruddy\CruddyController@index');
-    Route::get('thumb', 'Kalnoy\Cruddy\CruddyController@thumb');
+    Route::get('/', 'Kalnoy\Cruddy\Controllers\CruddyController@index');
+    Route::get('thumb', 'Kalnoy\Cruddy\Controllers\CruddyController@thumb');
 
     $entityPattern = app('cruddy.repository')->available();
     $entityPattern = "({$entityPattern})";
@@ -50,21 +50,21 @@ Route::group(['prefix' => Config::get('cruddy::uri')], function ()
     Route::get('{entity}',
     [
         'as' => 'cruddy.index', 
-        'uses' => 'Kalnoy\Cruddy\CruddyController@show'
+        'uses' => 'Kalnoy\Cruddy\Controllers\CruddyController@show'
     ])
     ->where('entity', $entityPattern);
 
     Route::get('{entity}/create',
     [
         'as' => 'cruddy.create',
-        'uses' => 'Kalnoy\Cruddy\CruddyController@show',
+        'uses' => 'Kalnoy\Cruddy\Controllers\CruddyController@show',
     ])
     ->where('entity', $entityPattern);
 
     Route::get('{entity}/{id}',
     [
         'as' => 'cruddy.show', 
-        'uses' => 'Kalnoy\Cruddy\CruddyController@show'
+        'uses' => 'Kalnoy\Cruddy\Controllers\CruddyController@show'
     ])
     ->where('entity', $entityPattern);
 });