+++ CHANGELOG.md
@@ -5,6 +5,8 @@ __0.3.0__
 *   [feature] Support Laravel 4.2
 *   [feature] A better date and time inputs
 *   [feature] A layout for forms
+*   [feature] Input addons
+*   [feature] Added `cruddy:compile` and `cruddy:clear-compiled` for compiling the schema
 *   [api] Removed `DateTime::format()`, the format is fixed now
 *   [api] Move lang-related stuff to `cruddy.lang` service
 *   [api] Move assets-related stuff to `cruddy.assets` service